Planet reworking continued: GPU-generated LOD (many bugs)

+++ stories/Eltanin/private/geo/celestial/planet.h
@@ -59,7 +59,7 @@ namespace eltanin::locality::planet {
         float spin; // radians around local +Y
         base::maybe<phys::Body::Id> well;
         geo::IcosaMap<std::int16_t> heights; // 0 = sea; ±reliefPeak maps to ±amplitude metres
-        geo::IcosaMap<std::uint16_t> covers; // two u8 facies: surface, then just below
+        geo::IcosaMap<std::uint32_t> covers; // four u8 facies, shallow to deep
 
     private:
         base::maybe<rmmr::scene::actor::PatchGrid::Id> shell;
